requirements.txt
OWSLib>=0.8.3
lxml>=3.2.1
cf_units>=0.1.1
requests>=2.2.1
isodate>=0.5.4
Jinja2>=2.7.3
setuptools>=15.0
pygeoif>=0.6
netCDF4>=1.3.0
regex==2017.07.28
pendulum>=1.2.4
PyYAML>=3.12
